The Punchline: Added implementation for new version update.
