+++ app/src/main/java/com/nextbigthing/thepunchline/ui/MainActivity.kt
@@ -8,18 +8,14 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.annotation.RequiresApi
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.window.DialogProperties
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.ValueEventListener
 import com.nextbigthing.thepunchline.navigation.Navigation
+import com.nextbigthing.thepunchline.ui.component.UpdateAppDialog
 import com.nextbigthing.thepunchline.ui.theme.ThePunchlineTheme
 import com.nextbigthing.thepunchline.util.JokesPreferenceHelper
 import com.nextbigthing.thepunchline.viewModel.JokesViewModel
@@ -31,37 +27,58 @@ import javax.inject.Inject
 class MainActivity : ComponentActivity() {
     @Inject lateinit var jokesPreferenceHelper: JokesPreferenceHelper
     @Inject lateinit var databaseReference: DatabaseReference
-    private val viewModel:JokesViewModel by viewModels()
+    private val viewModel: JokesViewModel by viewModels()
 
     @RequiresApi(Build.VERSION_CODES.S)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         jokesPreferenceHelper.clearPreferences()
+
+        // Get the version code of the app
         val packageManager = packageManager.getPackageInfo(packageName, 0)
         val versionCode = packageManager.versionCode
         jokesPreferenceHelper.saveInt("versionCode", versionCode)
+
+        // MutableState to control the dialog visibility
+        val showUpdateDialog = mutableStateOf(false)
+
         setContent {
-            fetchAppId()
             ThePunchlineTheme {
+                // Fetch the AppId when the activity starts
+                fetchAppId(showUpdateDialog)
+
+                // Display your navigation
                 Navigation(viewModel = viewModel, jokesPreferenceHelper = jokesPreferenceHelper)
+
+                // Show the Update Dialog if necessary
+                if (showUpdateDialog.value) {
+                    UpdateAppDialog(
+                        onUpdateClick = {
+                            Toast.makeText(this@MainActivity, "Updating...", Toast.LENGTH_SHORT).show()
+                            // Perform update action here
+                            showUpdateDialog.value = false
+                        },
+                        onDismiss = {
+                            showUpdateDialog.value = false
+                        }
+                    )
+                }
             }
         }
     }
 
-    private fun fetchAppId() {
+    private fun fetchAppId(showUpdateDialog: MutableState<Boolean>) {
         databaseReference.child("AppId").addListenerForSingleValueEvent(object : ValueEventListener {
             override fun onDataChange(snapshot: DataSnapshot) {
                 if (snapshot.exists()) {
                     // Fetch the AppId value
                     val versionCode = jokesPreferenceHelper.getInt("versionCode", 0)
                     val appId = snapshot.getValue(Int::class.java)
-                    if (appId != null) {
-                        if (appId>versionCode){
-
-                        }
+                    if (appId != null && appId > versionCode) {
+                        // Trigger the dialog if AppId is greater than the version code
+                        showUpdateDialog.value = true
                     }
                     Log.d("Firebase", "AppId: $appId")
-                    // Display or use the AppId value as needed
                 } else {
                     Log.e("Firebase", "AppId not found")
                 }
